requirements.txt
streamlit
openai-whisper
transformers
spacy
torch
smtplib
python-dotenv